cmds/delete: delete-destroy union function
Now the delete function will:
- stop and delete containers specified
- delete configurations
- delete shared folders
- prompt questions before deleting one or more containers

One thing changed was the way which this configurations/folders are removed.
Before we used the fs.unlink() function, but this function can't work
with folders (empty and non-empty).
So for convenience, we remove the folder through a simple execSync("rm -rf").
It may not be quite elegant but it's the more convenient way for now.
For removing just files, fs.unlinkSync is prefered.