completions: move completions generation
Now we follow the same mechanism for integrating command completion as
this guide https://cli.github.com/manual/gh_completion

So by calling:
- lxce completion -s {bash, zsh}
the command will output the corresponding script completion.

This decision has been mainly for the incovenien of running some lxce commands
as root and other not, so it up to the user to select and set up the corresponding
script in their setup.